--- .env.sample
@@ -1,5 +1,0 @@
-# https://apilayer.com/marketplace/exchangerates_data-api (for currencies)
-# http://api.marketstack.com/v1/eod?access_key={API_KEY_STOCK}&symbols={stock_name} (for stocks)
-
-API_KEY=place_your_API_KEY_here_for_currencies
-API_KEY_STOCK=place_your_API_KEY_here_for_stocks
