+++ src/constants.py
@@ -1,3 +1,4 @@
+import os
 from pathlib import Path
 
 BASE_DIR = str(Path(__file__).parent.parent)
@@ -5,6 +6,7 @@ DATA_DIR = BASE_DIR + '\\data\\'
 LOGS = BASE_DIR + '\\logs\\logs.log'
 
 FILE_NAME_API = 'VACANCIES_API'
+PATH_TO_VACANCIES_FILE = os.path.join(DATA_DIR, FILE_NAME_API)
 
 USER_MENU_LIST = [
     '1. Отфильтровать вакансии.',
